Added some network error catching

+++ src/gui/addon.cpp
@@ -37,21 +37,22 @@ Addon::Addon(const QString &name, QNetworkAccessManager *nam, QSettings *setting
 
 void Addon::checkForUpdate() {
     QNetworkReply *reply = (QNetworkReply*)QObject::sender();
-    pendingMd5 = reply->readLine(reply->size()).trimmed();
-    QByteArray reqPlayerVer = reply->readLine(reply->size()).trimmed();
+    if (reply->error() == QNetworkReply::NoError) {
+        pendingMd5 = reply->readLine(reply->size()).trimmed();
+        QByteArray reqPlayerVer = reply->readLine(reply->size()).trimmed();
 
-    if (MainWindow::versionNumberGreater(reqPlayerVer, QCoreApplication::instance()->applicationVersion())) {
-        QLabel *label = new QLabel(QString::fromUtf8("Väntande uppdatering av detta tillägg är inkompatibelt med din version av Pirateplayer. Vänligen <a href=\"http://pirateplay.se/player.html#install\">uppdatera</a>."), this);
-        label->setOpenExternalLinks(true);
-        layout()->addWidget(label);
-        load();
+        if (MainWindow::versionNumberGreater(reqPlayerVer, QCoreApplication::instance()->applicationVersion())) {
+            QLabel *label = new QLabel(QString::fromUtf8("Väntande uppdatering av detta tillägg är inkompatibelt med din version av Pirateplayer. Vänligen <a href=\"http://pirateplay.se/player.html#install\">uppdatera</a>."), this);
+            label->setOpenExternalLinks(true);
+            layout()->addWidget(label);
+        }
+        else if (pendingMd5 != settings->value("addon_" + name + "/md5sum", "0").toByteArray()) {
+            download();
+            return;
+        }
     }
-    else if (pendingMd5 != settings->value("addon_" + name + "/md5sum", "0").toByteArray()) {
-        download();
-    }
-    else {
-        load();
-    }
+
+    load();
 }
 
 void Addon::download() {
@@ -59,40 +60,52 @@ void Addon::download() {
     connect(reply, SIGNAL(finished()), SLOT(upgrade()));
 }
 
+//TODO: Check that this really works
 void Addon::upgrade() {
-    QByteArray *data = new QByteArray(((QIODevice*)QObject::sender())->readAll());
-    ArchiveExtractor *extractor = new ArchiveExtractor(this);
+    //QByteArray *data = new QByteArray(((QIODevice*)QObject::sender())->readAll());
+    QNetworkReply *reply = (QNetworkReply*)QObject::sender();
 
-    if (QCryptographicHash::hash(*data, QCryptographicHash::Md5).toHex() != pendingMd5) {
+    if (reply->error() == QNetworkReply::NoError) {
+        QByteArray *data = new QByteArray((reply->readAll()));
+        ArchiveExtractor *extractor = new ArchiveExtractor(this);
+
+        if (QCryptographicHash::hash(*data, QCryptographicHash::Md5).toHex() != pendingMd5) {
+            load();
+            return;
+        }
+        settings->setValue("addon_" + name + "/md5sum", pendingMd5);
+
+        removeDir(dir);
+        extractor->extractToDir(data, dir);
+        connect(extractor, SIGNAL(finished()), SLOT(load()));
+    } else {
         load();
-        return;
     }
-    settings->setValue("addon_" + name + "/md5sum", pendingMd5);
-
-    removeDir(dir);
-    extractor->extractToDir(data, dir);
-    connect(extractor, SIGNAL(finished()), SLOT(load()));
 }
 
 void Addon::load() {
-    QDeclarativeView *view = new QDeclarativeView(this);
-    QDeclarativeEngine *engine = view->engine();
-    QDeclarativeContext *rootContext = engine->rootContext();
+    QString mainQml = this->dir + "/main.qml";
 
-    engine->setNetworkAccessManagerFactory(new NetworkAccessManagerFactory);
+    if (QFile::exists(mainQml)) {
+        QDeclarativeView *view = new QDeclarativeView(this);
+        QDeclarativeEngine *engine = view->engine();
+        QDeclarativeContext *rootContext = engine->rootContext();
 
-    QString cwd = !this->dir.startsWith('/') ? this->dir = "/" + this->dir : this->dir;
-    rootContext->setContextProperty("cwd", cwd);
-    rootContext->setContextProperty("mainWindow", this->mainWindow);
+        engine->setNetworkAccessManagerFactory(new NetworkAccessManagerFactory);
 
-    view->setResizeMode(QDeclarativeView::SizeRootObjectToView);
-    view->setStyleSheet("background-color:transparent;");
-    view->setSource(QUrl::fromLocalFile(this->dir + "/main.qml"));
+        QString cwd = !this->dir.startsWith('/') ? this->dir = "/" + this->dir : this->dir;
+        rootContext->setContextProperty("cwd", cwd);
+        rootContext->setContextProperty("mainWindow", this->mainWindow);
 
-    QObject *rootItem = view->rootObject();
-    emit titleSet(rootItem->property("title").toString());
+        view->setResizeMode(QDeclarativeView::SizeRootObjectToView);
+        view->setStyleSheet("background-color:transparent;");
+        view->setSource(QUrl::fromLocalFile(mainQml));
 
-    ((QVBoxLayout*)layout())->addWidget(view, 1);
+        QObject *rootItem = view->rootObject();
+        emit titleSet(rootItem->property("title").toString());
+
+        ((QVBoxLayout*)layout())->addWidget(view, 1);
+    }
 }
 
 bool Addon::removeDir(const QString &dirName)
